feat: botão de copiar versículo
